+++ packages/fabric/media-picker/popup/src/components/views/upload/upload.tsx
@@ -409,7 +409,6 @@ const mapDispatchToProps = (
           id,
           mimeType,
           name,
-          parentId: '',
           size,
         },
         serviceName,