+++ packages/fabric/media-picker/popup/src/tools/fetcher/__tests__/fetcher.spec.tsx
@@ -1,8 +1,14 @@
 jest.mock('../../../../../src/util/getPreview');
 
+import axios from 'axios';
 import { Service } from '../../../domain';
 import { isImagePreview } from '../../../../../src/domain/preview';
-import { flattenAccounts, MediaApiFetcher } from '../fetcher';
+import {
+  MediaApiFetcher,
+  flattenAccounts,
+  GiphyResponse,
+  GiphyImage,
+} from '../fetcher';
 import getPreviewFromBlob from '../../../../../src/util/getPreview';
 
 describe('Fetcher', () => {
@@ -26,6 +32,10 @@ describe('Fetcher', () => {
     querySpy = jest.fn();
 
     querySpy.mockReturnValue(Promise.resolve({}));
+  });
+
+  afterEach(() => {
+    (axios.request as any).mockReset();
   });
 
   describe('getPreview()', () => {
@@ -111,6 +121,157 @@ describe('Fetcher', () => {
           type: 'dropbox',
         },
       ]);
+    });
+  });
+
+  describe('GIPHY methods', () => {
+    const gifId = 'some-gif-id';
+    const gifSlug = `file-slug-${gifId}`;
+    const gifUrl = 'some-gif-url';
+    const gifFileSizeStr = '1234567';
+    const gifFileSizeNum = 1234567;
+    const gifWidth = '200';
+    const gifHeight = '363';
+
+    const originalGifUrl = 'some-original-gif-url';
+
+    const images = {
+      fixed_width: {
+        url: gifUrl,
+        size: gifFileSizeStr,
+        width: gifWidth,
+        height: gifHeight,
+      } as GiphyImage,
+      original: {
+        url: originalGifUrl,
+      } as GiphyImage,
+    };
+
+    const response: GiphyResponse = {
+      data: [{ id: gifId, slug: gifSlug, images }],
+      pagination: {
+        total_count: 100,
+        count: 25,
+        offset: 0,
+      },
+    };
+
+    describe('fetchTrendingGifs()', () => {
+      it('ignore the offset when it is 0', () => {
+        const offset = 0;
+        const fetcher = new MediaApiFetcher();
+
+        fetcher.fetchTrendingGifs(offset);
+        expect(axios.request).toHaveBeenCalledTimes(1);
+        expect(
+          (axios.request as any).mock.calls[0][0].url.indexOf(
+            `&offset=${offset}`,
+          ),
+        ).toBe(-1);
+      });
+
+      it('appends passed in offset to the query string when it is greater than 0', () => {
+        const offset = 25;
+        const fetcher = new MediaApiFetcher();
+
+        fetcher.fetchTrendingGifs(offset);
+        expect(axios.request).toHaveBeenCalledTimes(1);
+        expect((axios.request as any).mock.calls[0][0].params.offset).toEqual(
+          offset,
+        );
+      });
+
+      it('maps the GiphyResponse into GiphyCardModels', async () => {
+        const fetcher = new MediaApiFetcher();
+        (axios.request as any).mockReturnValue({ data: response });
+
+        const result = await fetcher.fetchTrendingGifs();
+        expect(axios.request).toHaveBeenCalledTimes(1);
+        expect(result).toEqual({
+          totalResultCount: 100,
+          cardModels: [
+            {
+              metadata: {
+                id: gifId,
+                name: 'file-slug',
+                mediaType: 'image',
+                size: gifFileSizeNum,
+              },
+              dataURI: gifUrl,
+              dimensions: {
+                width: 200,
+                height: 363,
+              },
+            },
+          ],
+        });
+      });
+    });
+
+    describe('fetchGifsRelevantToSearch()', () => {
+      it('ignore the offset when it is 0', () => {
+        const queryString = 'some-gif-search';
+        const offset = 0;
+        const fetcher = new MediaApiFetcher();
+
+        fetcher.fetchGifsRelevantToSearch(queryString, offset);
+        expect(axios.request).toHaveBeenCalledTimes(1);
+        expect(
+          (axios.request as any).mock.calls[0][0].url.indexOf(
+            `&offset=${offset}`,
+          ),
+        ).toBe(-1);
+      });
+
+      it('appends passed in query string to the queried url', () => {
+        const queryString = 'some-gif-search';
+        const fetcher = new MediaApiFetcher();
+
+        fetcher.fetchGifsRelevantToSearch(queryString);
+        expect(axios.request).toHaveBeenCalledTimes(1);
+        expect((axios.request as any).mock.calls[0][0].params.q).toEqual(
+          queryString,
+        );
+      });
+
+      it('appends passed in offset to the query string when it is greater than 0', () => {
+        const queryString = 'some-gif-search';
+        const offset = 25;
+        const fetcher = new MediaApiFetcher();
+
+        fetcher.fetchGifsRelevantToSearch(queryString, offset);
+        expect(axios.request).toHaveBeenCalledTimes(1);
+        expect((axios.request as any).mock.calls[0][0].params.offset).toEqual(
+          offset,
+        );
+      });
+
+      it('maps the GiphyResponse into GiphyCardModels', async () => {
+        const queryString = 'some-gif-search';
+        const fetcher = new MediaApiFetcher();
+        (axios.request as any).mockReturnValue({ data: response });
+
+        const result = await fetcher.fetchGifsRelevantToSearch(queryString);
+        expect(axios.request).toHaveBeenCalledTimes(1);
+        expect(result).toEqual({
+          totalResultCount: 100,
+          cardModels: [
+            {
+              metadata: {
+                id: gifId,
+                name: 'file-slug',
+                mediaType: 'image',
+                size: gifFileSizeNum,
+              },
+              dataURI: gifUrl,
+              dimensions: {
+                width: 200,
+                height: 363,
+              },
+            },
+          ],
+        });
+      });
     });
   });
 });